requirements.txt
streamlit
google-generativeai>=0.8.3
openai>=1.0.0
Pillow
watchdog
PyExifTool
opencv-python-headless
pandas
openpyxl
python-dotenv
streamlit-option-menu
typing_extensions